requirements.txt
flask>=3.0
flask-cors>=4.0
anthropic>=0.34
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
python-dotenv>=1.0
requests>=2.31
pdfplumber>=0.11
APScheduler>=3.10
pytz>=2024.1
openai>=1.0
google-generativeai>=0.8
# Production / multi-user additions
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
PyJWT>=2.8
cryptography>=42
boto3>=1.34
gunicorn>=22
